Fix conflicts with others' code
