+++ routes/shows.js
@@ -54,20 +54,20 @@ router.get('/:id', async(req, res) => {
             show.summary = "NA";
         }
 
-        res.render('shows/single', {
+        res.render('movies/single', {
             show: show,
             title: show.name
         });
     } catch (e) {
         if (e.response && e.response.status && e.response.status === 404) {
-            res.status(404).render('shows/error', {
-                title: "No TV Show Found",
+            res.status(404).render('movies/error', {
+                title: "No Movie Found",
                 hasErrors: true,
-                error: "No TV Show Found with TV Show ID = `" + Number(req.params.id) + "`"
+                error: "No Movie Found with Movie ID = `" + Number(req.params.id) + "`"
             });
         } else {
-            res.status(404).render('shows/error', {
-                title: "No TV Show Found",
+            res.status(404).render('movies/error', {
+                title: "No Movie Found",
                 hasErrors: true,
                 error: e
             });
