--- views/shows/single.handlebars
@@ -1,24 +1,0 @@
-<main class="movie">
-  <div>
-    <h1>{{movie.Title}}</h1>
-    <img alt="{{movie.Title}}" src="{{movie.Poster}}" />
-    <dl>
-      <dt>Language</dt>
-      <dd>{{movie.Language}}</dd>
-      <dt>Genres</dt>
-      <dd>{{movie.Genre}}</dd>
-      <dt>Average Rating</dt>
-      {{#each movie.Ratings}}
-      <dd>
-        <ul>
-          <li>{{this}}</li>
-        </ul>
-      </dd>
-      {{/each}} 
-      <dt>Runtime</dt>
-      <dd>{{movie.Runtime}}</dd>
-      <dt>Plot</dt>
-      <dd>{{movie.Plot}}</dd>
-    </dl>
-  </div>
-</main>
