+++ views/users/profile.handlebars
@@ -10,7 +10,7 @@
             {{#each watchList}}
             <td class='thumbnail-list-td'>
                 {{!-- TODO Fix links after movie page is ready --}}                
-                <a href='/movie/{{this._id}}'>
+                <a href='/movies/{{this._id}}'>
                     <div class='thumbnail-item-title'>{{ this.name }}</div>
                     <img class='thumbnail-item-img' alt="Poster for {{movie.name}}" src={{this.poster}} />
                 </a>
@@ -48,7 +48,7 @@
         <table>
             {{#each likedMovies}}
             <td class='thumbnail-list-td'>
-                <a href='/movie/{{this._id}}'>
+                <a href='/movies/{{this._id}}'>
                     <div class='thumbnail-item-title'>{{ this.name }}</div>
                     <img class='thumbnail-item-img' alt="Poster for {{movie.name}}" src={{this.poster}} />
                 </a>
@@ -116,7 +116,7 @@
         <table>
             {{#each dislikedMovies}}
             <td class='thumbnail-list-td'>
-                <a href='/movie/{{this._id}}'>
+                <a href='/movies/{{this._id}}'>
                     <div class='thumbnail-item-title'>{{ this.name }}</div>
                     <img class='thumbnail-item-img' alt="Poster for {{movie.name}}" src={{this.poster}} />
                 </a>
